+++ vakt/storage/mongo.py
@@ -3,9 +3,11 @@ MongoDB Storage and Migrations for Policies.
 """
 
 import logging
+import copy
 
 import bson.json_util as b_json
 from pymongo.errors import DuplicateKeyError
+import jsonpickle.tags
 
 from ..storage.abc import Storage, Migration
 from ..exceptions import PolicyExistsError, UnknownCheckerType
@@ -128,9 +130,13 @@ class MongoStorage(Storage):
             yield self.__prepare_from_doc(doc)
 
 
-class Migration0To1x0x3(Migration):
+##############
+# Migrations #
+##############
+
+class Migration0To1x1x0(Migration):
     """
-    Migration between versions 0 and 1.0.3
+    Migration between versions 0 and 1.1.0
     """
 
     def __init__(self, storage):
@@ -155,3 +161,85 @@ class Migration0To1x0x3(Migration):
     def down(self):
         for field in self.multi_key_indices:
             self.storage.collection.drop_index(self.index_name(field))
+
+
+class Migration1x1x0To1x1x1(Migration):
+    """
+    Migration between versions 1.1.0 and 1.1.1
+    """
+
+    class Irreversible(Exception):
+        pass
+
+    def __init__(self, storage):
+        self.storage = storage
+        self._type_marker = jsonpickle.tags.OBJECT
+
+    @property
+    def order(self):
+        return 2
+
+    def up(self):
+        def process(doc):
+            doc_to_save = copy.deepcopy(doc)
+            rules_to_save = {}
+            for name, rule_str in doc['rules'].items():
+                rule = b_json.loads(rule_str)
+                rule_to_save = {self._type_marker: rule['type']}
+                rule_to_save.update(rule['contents'])
+                rules_to_save[name] = rule_to_save
+            doc_to_save['rules'] = rules_to_save
+            return doc_to_save
+        self._docs_iteration(processor=process)
+
+    def down(self):
+        def process(doc):
+            doc_to_save = copy.deepcopy(doc)
+            rules_to_save = {}
+            for name, rule in doc['rules'].items():
+                rule_type = rule[self._type_marker]
+                rule_contents = rule.copy()
+                del rule_contents[self._type_marker]
+                rule_to_save = {'type': rule_type, 'contents': {}}
+                # check if we are dealing with 3-rd party or custom rules
+                if not rule_type.startswith('vakt.rules.'):
+                    for value in rule_contents.values():
+                        # if rule has non-primitive data as its contents - we can't revert it to 1.1.0
+                        if isinstance(value, dict) and jsonpickle.tags.RESERVED.intersection(value.keys()):
+                            raise self.Irreversible
+                # vakt's own RegexMatchRule couldn't be stored in mongo because is has non-primitive data,
+                # so it's impossible to put it to storage if we revert time back to 1.1.0
+                elif rule_type == 'vakt.rules.string.RegexMatchRule':
+                    raise self.Irreversible
+                rule_to_save['contents'].update(rule_contents)
+                rules_to_save[name] = b_json.dumps(rule_to_save)
+            # report or save document
+            doc_to_save['rules'] = rules_to_save
+            return doc_to_save
+        self._docs_iteration(processor=process)
+
+    def _docs_iteration(self, processor):
+        failed_policies = []
+        cur = self.storage.collection.find()
+        for doc in cur:
+            try:
+                log.info('Trying to migrate Policy with UID: %s' % doc['uid'])
+                new_doc = processor(doc)
+                self.storage.collection.update_one({'_id': new_doc['uid']}, {"$set": new_doc}, upsert=False)
+                log.info('Policy with UID was migrated: %s' % doc['uid'])
+            except self.Irreversible:
+                log.warning('(Probably) custom Policy is irreversible: %s', doc)
+                failed_policies.append(doc)
+            except Exception as e:
+                log.exception('Unexpected exception occurred while migrating Policy: %s', doc)
+                failed_policies.append(doc)
+        if failed_policies:
+            msg = "\n".join([
+                'Migration was unable to convert some Policies.',
+                'But they were left in the database as-is.',
+                'Some of them are custom ones, some are just malformed JSON docs.',
+                'You must convert them manually or delete them. See above log output for details of migration',
+                'Policies Mongo IDs are:',
+                str([p['_id'] for p in failed_policies])
+            ])
+            log.error(msg)